feat: use new API for node allocation & type validation via types env
Now using our new API to allocate our nodes, add them as children, etc. We now use an env, a types env from types env, and we can free them once the result is done with var_decl, etc., so we can get types and return errors if they occur.

For example, it's supposed to be:
a : integer = 69
But if we change integer to int like:
a : int = 69
We get:
ERROR: Mismatched types